pyproject.toml
[tool.poetry]
name = "portfolio-backend"
version = "0.1.0"
description = ""
authors = ["corey-marchand <corey.marchand@me.com>"]

[tool.poetry.dependencies]
python = "^3.8"
django = "^3.1.1"
gunicorn = "^20.0.4"
whitenoise = "^5.2.0"
django-cors-headers = "^3.5.0"
psycopg2-binary = "^2.8.6"
django-environ = "^0.4.5"
djangorestframework = "^3.12.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
